Totals are one page. Adds print style. Starts Edit
Changes the way the Totals page creates the table, so
it will now print to a single sheet of paper. Adds a
print style for further control over printing process.
Also added the edit order page. Goal is to have a place
for users to look up students and alter order. That page
needs a lot of help yet.