Adds year table and students_report
Year table tracks each year, its student table, and the profits
and prices for fruit that year. Added relevant details to data_design.
First page to start using it is the newly added students_report.
Students_report, once it's working, will show the check totals and
profits for each student. It is not quite working correctly yet, but
it will be soon.